Move JSON Test Format To integration-testing
